+++ src/lib/UI/pageComponents/other/DeleteAllDataBlock.svelte
@@ -35,7 +35,6 @@
     @apply text-red-700 border border-red-700 font-medium rounded-lg text-sm px-5 py-2.5 text-center;
     @apply hover:text-white hover:bg-red-800 focus:ring-red-300 focus:ring-4 focus:outline-none;
     @apply dark:border-red-500 dark:text-red-500 dark:hover:text-white dark:hover:bg-red-600 dark:focus:ring-red-900;
-
     margin-left: 1rem;
   }
 </style>
