+++ src/lib/UI/pageComponents/other/ExportDataBlock.svelte
@@ -38,7 +38,6 @@
   button {
     @apply text-green-500 border border-green-500 font-medium rounded-lg text-sm px-5 py-2.5 text-center;
     @apply hover:text-white hover:bg-green-400 focus:ring-green-300 focus:ring-4 focus:outline-none;
-
     margin-left: 1rem;
   }
 </style>
