+++ src/lib/UI/pageComponents/other/ImportDataBlock.svelte
@@ -60,7 +60,6 @@
   button {
     @apply text-yellow-500 border border-yellow-500 font-medium rounded-lg text-sm px-5 py-2.5 text-center;
     @apply hover:text-white hover:bg-yellow-400 focus:ring-yellow-300 focus:ring-4 focus:outline-none;
-
     margin-left: 1rem;
   }
 </style>
