+++ src/routes/recipes/[id=integer]/+page.svelte
@@ -289,7 +289,6 @@
 
   .my-button {
     @apply bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded;
-
     width: 100%;
     margin-bottom: 16px;
   }
